backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
psycopg[binary]>=3.1.0
python-dotenv==1.0.0
reportlab==4.0.9
Pillow>=10.0.0
